chore: hook up new way of finding adds

+++ Kitsu/Controllers/LibraryController.cs
@@ -91,7 +91,7 @@ namespace Kitsu.Controllers
             return rtn;
         }
 
-        public async Task<List<LibraryEntry>> GetLibraryCollectionByIdsAsync(int userId, int[] ids, LibraryType type = LibraryType.All, LibraryStatus status = LibraryStatus.All)
+        public async Task<List<LibraryEntry>> GetLibraryCollectionByIdsAsync(int userId, long[] ids, LibraryType type = LibraryType.All, LibraryStatus status = LibraryStatus.All)
         {
             var rtn = new List<LibraryEntry>();
